requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-3-jet==1.0.8
dnspython==2.0.0
filelock==3.0.12
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py3-validate-email==0.2.10
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
